refactor: 폴더 구조 변경 및 가독성을 위한 공백 라인 추가
