+++ src/constants/index.js
@@ -24,10 +24,13 @@ export {
   MIN_OF_RANDOM,
   MAX_OF_RANDOM,
   MOVE_CONDITION_NUMBER,
+
   // validation
   MIN_NUMBER_OF_CARS,
   MIN_LENGTH_OF_CAR_NAME,
   MAX_LENGTH_OF_CAR_NAME,
   MIN_OF_RACING_COUNT,
+
+  // Error Message
   ERROR_MESSAGE,
 };
